# Fix `matplotlib.plot_contour` bug