--- a/optuna/visualization/matplotlib/_contour.py
+++ b/optuna/visualization/matplotlib/_contour.py
@@ -311,35 +311,36 @@ def _calculate_griddata(
     xi = np.array([])
     yi = np.array([])
     zi = np.array([])
+
+    if _is_log_scale(trials, x_param):
+        padding_x = (np.log10(x_values_max) - np.log10(x_values_min)) * AXES_PADDING_RATIO
+        x_values_min = np.power(np.log10(x_values_min) - padding_x, 10)
+        x_values_max = np.power(np.log10(x_values_max) + padding_x, 10)
+        xi = np.logspace(np.log10(x_values_min), np.log10(x_values_max), contour_point_num)
+        x_array = np.log10(x_values)
+    else:
+        padding_x = (x_values_max - x_values_min) * AXES_PADDING_RATIO
+        x_values_min -= padding_x
+        x_values_max += padding_x
+        xi = np.linspace(x_values_min, x_values_max, contour_point_num)
+        x_array = np.array(x_values)
+
+    if _is_log_scale(trials, y_param):
+        padding_y = (np.log10(y_values_max) - np.log10(y_values_min)) * AXES_PADDING_RATIO
+        y_values_min = np.power(np.log10(y_values_min) - padding_y, 10)
+        y_values_max = np.power(np.log10(y_values_max) + padding_y, 10)
+        yi = np.logspace(np.log10(y_values_min), np.log10(y_values_max), contour_point_num)
+        y_array = np.log10(y_values)
+    else:
+        padding_y = (y_values_max - y_values_min) * AXES_PADDING_RATIO
+        y_values_min -= padding_y
+        y_values_max += padding_y
+        yi = np.linspace(y_values_min, y_values_max, contour_point_num)
+        y_array = np.array(y_values)
+
+    # create irregularly spaced map of trial values
+    # and interpolate it with Plotly algorithm
     if x_param != y_param:
-        if _is_log_scale(trials, x_param):
-            padding_x = (np.log10(x_values_max) - np.log10(x_values_min)) * AXES_PADDING_RATIO
-            x_values_min = np.power(np.log10(x_values_min) - padding_x, 10)
-            x_values_max = np.power(np.log10(x_values_max) + padding_x, 10)
-            xi = np.logspace(np.log10(x_values_min), np.log10(x_values_max), contour_point_num)
-            x_array = np.log10(x_values)
-        else:
-            padding_x = (x_values_max - x_values_min) * AXES_PADDING_RATIO
-            x_values_min -= padding_x
-            x_values_max += padding_x
-            xi = np.linspace(x_values_min, x_values_max, contour_point_num)
-            x_array = np.array(x_values)
-
-        if _is_log_scale(trials, y_param):
-            padding_y = (np.log10(y_values_max) - np.log10(y_values_min)) * AXES_PADDING_RATIO
-            y_values_min = np.power(np.log10(y_values_min) - padding_y, 10)
-            y_values_max = np.power(np.log10(y_values_max) + padding_y, 10)
-            yi = np.logspace(np.log10(y_values_min), np.log10(y_values_max), contour_point_num)
-            y_array = np.log10(y_values)
-        else:
-            padding_y = (y_values_max - y_values_min) * AXES_PADDING_RATIO
-            y_values_min -= padding_y
-            y_values_max += padding_y
-            yi = np.linspace(y_values_min, y_values_max, contour_point_num)
-            y_array = np.array(y_values)
-
-        # create irregularly spaced map of trial values
-        # and interpolate it with Plotly algorithm
         zmap = _create_zmap(x_array, y_array, z_values, xi, yi)
         _interpolate_zmap(zmap, contour_point_num)
         zi = _create_zmatrix_from_zmap(zmap, contour_point_num)