Close #3039  about `matplotlib.plot_contour` bug 

## Description

When the top left subplot is being plotted, the `len(zi) == 0` thus
- `ax.set_xlim(x_values_range[0], x_values_range[1])`, and
- `ax.set_ylim(y_values_range[0], y_values_range[1])`
are not executed and resulting the top left subplot doesn't have proper `xlim` and `ylim`