@nzw0301 2021-10-31
I took a benchmark to compare the speed between the master branch and https://github.com/optuna/optuna/pull/3046#issuecomment-954904185, namely, not-skipped using the same code as in https://github.com/optuna/optuna/pull/3046#issuecomment-954572745. 

- Master: 14.2 s ± 620 ms per loop (mean ± std. dev. of 7 runs, 1 loop each) 
- not-skipped: 21.5 s ± 723 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

This difference might not be acceptable...