@IEP 2021-11-01
The heavy part that causes significant addition to the execution time is related to

```python
zmap = _create_zmap(x_array, y_array, z_values, xi, yi)
_interpolate_zmap(zmap, contour_point_num)
zi = _create_zmatrix_from_zmap(zmap, contour_point_num)
```

and the condition `x_param != y_param` can be moved to guard this snippet.

Benchmark:
- Original: 9.05 s ± 59.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- Patch (move condition): 9.03 s ± 178 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)